Add AllTheProviders component for improved test rendering and refactor test-utils to utilize it
