+++ webapp/src/test/utils/test-utils.tsx
@@ -1,8 +1,7 @@
-/* eslint-disable react-refresh/only-export-components */
-import { type ReactElement, type ReactNode } from 'react';
+import { type ReactElement } from 'react';
 import { render, type RenderOptions, screen, fireEvent, waitFor } from '@testing-library/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { MemoryRouter } from 'react-router-dom';
+import { QueryClient } from '@tanstack/react-query';
+import { AllTheProviders } from './TestWrapper';
 
 /**
  * Create a new QueryClient for each test to ensure isolation.
@@ -20,29 +19,6 @@ export function createTestQueryClient(): QueryClient {
       },
     },
   });
-}
-
-/**
- * Internal helper component.
- */
-interface AllTheProvidersProps {
-  children: ReactNode;
-  queryClient?: QueryClient;
-  initialRoutes?: string[];
-}
-
-function AllTheProviders({
-  children,
-  queryClient = createTestQueryClient(),
-  initialRoutes = ['/'],
-}: AllTheProvidersProps) {
-  return (
-    <QueryClientProvider client={queryClient}>
-      <MemoryRouter initialEntries={initialRoutes}>
-        {children}
-      </MemoryRouter>
-    </QueryClientProvider>
-  );
 }
 
 /**
